+++ interface/Vue/src/utils/cookieHelper.ts
@@ -1,0 +1,16 @@
+export function setCookie(name: string, value: string, days = 7) {
+  let expires = '';
+  if (days) {
+    const date = new Date();
+    date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000);
+    expires = '; expires=' + date.toUTCString();
+  }
+  document.cookie = name + '=' + (value || '') + expires + '; path=/; SameSite=Lax';
+}
+export function eraseCookie(name: string) {
+  document.cookie = name + '=; Max-Age=-99999999; path=/;';
+}
+export function getCookie(name: string): string | null {
+  const match = document.cookie.match(new RegExp('(^| )' + name + '=([^;]+)'));
+  return match ? match[2] : null;
+}
